@@ -17,7 +17,7 @@ import {
 import { UserService } from 'src/app/core/user.service';
 import { ErrorService } from 'src/app/core/error.service';
 import { SplitService } from 'src/app/core/split.service';
-import { DashboardService } from '../dashboard.service';
+import { DashboardService } from 'src/app/dashboard/dashboard.service';
 import { throwError } from 'rxjs';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
@@ -26,6 +26,7 @@ import { SidenavDrawerService } from 'src/app/core/sidenav-drawer.service';
 import { By } from '@angular/platform-browser';
 import { StationWidgetComponent } from '../widgets/station-widget/station-widget.component';
 import { GridsterModule } from 'angular-gridster2';
+import { DashboardData, WidgetType } from 'src/models';
 
 describe('DashboardComponent', () => {
   let component: DashboardComponent;
@@ -97,6 +98,11 @@ describe('DashboardComponent', () => {
 
   it('should show error message when request for station widgets', () => {
     component.viewNewDashboard = true;
+    component.dashboardData = {
+      widgets: [],
+      name: 'General',
+      rithmId: '',
+    };
     spyOn(
       TestBed.inject(DashboardService),
       'getDashboardWidgets'
@@ -137,5 +143,103 @@ describe('DashboardComponent', () => {
     const menuBtn = fixture.debugElement.query(By.css('#menu-button'));
     menuBtn.triggerEventHandler('click', null);
     expect(spy).toHaveBeenCalled();
+  });
+
+  it('should call service from generateNewPersonalDashboard', () => {
+    const spyService = spyOn(
+      TestBed.inject(DashboardService),
+      'generateNewPersonalDashboard'
+    ).and.callThrough();
+
+    component.generateNewPersonalDashboard();
+    expect(spyService).toHaveBeenCalled();
+  });
+
+  it('should show an error message when calling generateNewPersonalDashboard', () => {
+    spyOn(
+      TestBed.inject(DashboardService),
+      'generateNewPersonalDashboard'
+    ).and.returnValue(
+      throwError(() => {
+        throw new Error();
+      })
+    );
+
+    const spyError = spyOn(
+      TestBed.inject(ErrorService),
+      'displayError'
+    ).and.callThrough();
+
+    component.generateNewPersonalDashboard();
+    expect(spyError).toHaveBeenCalled();
+  });
+
+  it('should call service to update a personal dashboard', () => {
+    component.viewNewDashboard = true;
+    const updatePersonalDashboardSpy = spyOn(
+      TestBed.inject(DashboardService),
+      'updatePersonalDashboard'
+    ).and.callThrough();
+
+    component['updatePersonalDashboard']();
+    expect(updatePersonalDashboardSpy).toHaveBeenCalled();
+  });
+
+  it('should call service from generateNewOrganizationDashboard', () => {
+    const spyService = spyOn(
+      TestBed.inject(DashboardService),
+      'generateNewOrganizationDashboard'
+    ).and.callThrough();
+
+    component.generateNewOrganizationDashboard();
+    expect(spyService).toHaveBeenCalled();
+  });
+
+  it('should show an error message when calling generateNewOrganizationDashboard', () => {
+    spyOn(
+      TestBed.inject(DashboardService),
+      'generateNewOrganizationDashboard'
+    ).and.returnValue(
+      throwError(() => {
+        throw new Error();
+      })
+    );
+
+    const spyError = spyOn(
+      TestBed.inject(ErrorService),
+      'displayError'
+    ).and.callThrough();
+
+    component.generateNewOrganizationDashboard();
+    expect(spyError).toHaveBeenCalled();
+  });
+
+  it('should call the service updateOrganizationDashboard ', () => {
+    const dashboardData: DashboardData = {
+      rithmId: '',
+      name: 'new name',
+      widgets: [
+        {
+          cols: 1,
+          rows: 2,
+          x: 0,
+          y: 0,
+          widgetType: WidgetType.Document,
+          data: 'string',
+          minItemRows: 1,
+          maxItemRows: 2,
+          minItemCols: 1,
+          maxItemCols: 2,
+        },
+      ],
+    };
+
+    const methodService = spyOn(
+      TestBed.inject(DashboardService),
+      'updateOrganizationDashboard'
+    ).and.callThrough();
+
+    component.updateOrganizationDashboard(dashboardData);
+    expect(methodService).toHaveBeenCalledWith(dashboardData);
   });
 });
