@@ -6,7 +6,7 @@ import {
 } from '@angular/core/testing';
 import { DocumentService } from 'src/app/core/document.service';
 import { StationDocumentsModalComponent } from './station-documents-modal.component';
-import { MockDocumentService } from 'src/mocks';
+import { MockDocumentService, MockUserService } from 'src/mocks';
 import { MockPopupService } from 'src/mocks';
 import { PopupService } from 'src/app/core/popup.service';
 import { StationDocumentsModalData, UserType } from 'src/models';
@@ -14,15 +14,19 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { MatTooltipHarness } from '@angular/material/tooltip/testing';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import {
+  NoopAnimationsModule,
+  BrowserAnimationsModule,
+} from '@angular/platform-browser/animations';
 import {
   MatDialogModule,
   MatDialogRef,
   MAT_DIALOG_DATA,
 } from '@angular/material/dialog';
 import { RouterTestingModule } from '@angular/router/testing';
-import { LoadingIndicatorComponent } from '../loading-indicator/loading-indicator.component';
+import { LoadingIndicatorComponent } from 'src/app/shared/loading-indicator/loading-indicator.component';
 import { MockComponent } from 'ng-mocks';
+import { UserService } from 'src/app/core/user.service';
 
 const DIALOG_TEST_DATA: StationDocumentsModalData = {
   stationName: 'A Station',
@@ -47,6 +51,7 @@ describe('StationDocumentsModalComponent', () => {
       ],
       imports: [
         RouterTestingModule,
+        BrowserAnimationsModule,
         NoopAnimationsModule,
         MatTooltipModule,
         MatDialogModule,
@@ -57,6 +62,7 @@ describe('StationDocumentsModalComponent', () => {
         { provide: MAT_DIALOG_DATA, useValue: DIALOG_TEST_DATA },
         { provide: DocumentService, useClass: MockDocumentService },
         { provide: PopupService, useClass: MockPopupService },
+        { provide: UserService, useClass: MockUserService },
       ],
     }).compileComponents();
   });
