@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { StationDocumentsModalComponent } from './station-documents-modal.component';
 import { MatTooltipModule } from '@angular/material/tooltip';
-import { UserAvatarModule } from '../user-avatar/user-avatar.module';
-import { LoadingIndicatorModule } from '../loading-indicator/loading-indicator.module';
-import { PaginationModule } from '../pagination/pagination.module';
+import { UserAvatarModule } from 'src/app/shared/user-avatar/user-avatar.module';
+import { LoadingIndicatorModule } from 'src/app/shared/loading-indicator/loading-indicator.module';
+import { PaginationModule } from 'src/app/shared/pagination/pagination.module';
 import { MatDialogModule } from '@angular/material/dialog';
 
 @NgModule({
